refactor(data-layer): conform migrated data file names to the mobile rules
Two names came into src/shared/data with the spellings they had as
package modules. Outside packages/**, a function or constant module is
camelCase.ts, and a leading underscore is reserved for router tokens:

- preference-schema.ts -> preferenceSchema.ts (kebab-case is the
  packages/** rule; preferenceTypes.ts already referred to the file by
  its camelCase name)
- _endpointHelpers.ts -> endpointHelpers.ts (the underscore was a
  desktop-sync marker, and data no longer mirrors desktop)
