+++ src/backend/data/services/utils/orderKey.ts
@@ -15,7 +15,7 @@ import { generateKeyBetween, generateNKeysBetween } from 'fractional-indexing';
 
 import { loggerService } from '@/shared/core/logger/LoggerService';
 import { DataApiErrorFactory } from '@/shared/data/api/errors';
-import type { OrderRequest } from '@/shared/data/api/schemas/_endpointHelpers';
+import type { OrderRequest } from '@/shared/data/api/schemas/endpointHelpers';
 
 type TxLike = any;
