+++ src/shared/data/preference/index.ts
@@ -9,8 +9,8 @@ export type {
   PermissionPreferenceKey,
   PreferenceKeyType,
   PreferenceSchema,
-} from './preference-schema';
-export { FONT_SIZE_STEPS, PreferenceDefaults } from './preference-schema';
+} from './preferenceSchema';
+export { FONT_SIZE_STEPS, PreferenceDefaults } from './preferenceSchema';
 export type { LanguageVarious, PermissionMode, PreferenceUpdateOptions } from './preferenceTypes';
 export { ThemeMode } from './preferenceTypes';
 export { getDefaultValue, getPreferenceKeys, isPreferenceKey } from './preferenceUtils';
